adapter/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
anthropic==0.49.0
pydantic==2.6.4
python-dotenv==1.0.1
